requirements.txt
# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Database
sqlalchemy[asyncio]>=2.0.0
aiosqlite>=0.19.0

# LLM Integration
litellm>=1.0.0

# Git Operations
gitpython>=3.1.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
aiofiles>=23.0.0
websockets>=12.0
python-multipart>=0.0.6
